refactor(core,proxy): move {model}:{profile} routing into the request pipeline
`resolve_route` decides whether a requested id names a model, a model plus
a configured profile, or a profile that does not exist. It lived in
`gglib-proxy`, so the proxy was the only surface that could select a
profile at all — not by design, but because that is where it was first
needed. The CLI is about to need it, and `gglib-core` is the only crate
both can reach.

It lands in `request_pipeline` rather than `domain` because it takes a
`&dyn ModelCatalogPort`, and `domain` is port-free — every `crate::ports`
reference under it is a doc link or an error type, never a trait object.
`request_pipeline::resolve` is the exact precedent: same catalog-round-trip
shape, same fail-open-on-CatalogError policy.

But it is not a *stage*: every stage of `apply()` shapes a request already
known to belong to some model, and `resolve()` cannot build a
`ModelContext` until the base name is known — which is what stripping the
suffix produces. So the README gains a "Routing" section ahead of
"Resolution" that writes that ordering down, rather than filing this under
a heading whose members all feed `ModelContext`.

`variant_entries` stays in the proxy: it needs `crate::models::ModelInfo`,
and advertising ids in `/v1/models` is a proxy concern in a way that
routing an id is not. `NamedCatalog` is deliberately duplicated rather than
moved, because the round-trip test that pins advertised ids resolving back
to their model is the one test that must span both crates.

Pure move. The nine routing tests travel with the code and pass in core;
`integration_inference_profiles.rs` is untouched and its 13 tests still
pass, which is the real proof that nothing about the behaviour changed.

Full pre-commit green.
